travis-ci: check cabal cache directory.

+++ travis-CI/cabal-hvr-ghc/install.sh
@@ -12,12 +12,14 @@ echo "$(ghc --version) [$(ghc --print-project-git-commit-id 2> /dev/null || echo
 
 ## In package list cache hit case, cabal install may fail
 (
-    cd $HOME/.cabal/packages/hackage.haskell.org/
-    rm -f \
-       00-index.tar.gz.etag \
-       00-index.tar.gz \
-       00-index.tar \
-       00-index.cache
+    if [ -d $HOME/.cabal/packages/hackage.haskell.org/ ]; then
+        cd $HOME/.cabal/packages/hackage.haskell.org/
+        rm -f \
+           00-index.tar.gz.etag \
+           00-index.tar.gz \
+           00-index.tar \
+           00-index.cache
+    fi
 )
 
 custom_retry cabal update -v
